+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [0.48.2](https://github.com/0xranx/golembot/compare/v0.48.1...v0.48.2) (2026-07-07)
+
+
+### Bug Fixes
+
+* **dingtalk:** download images and files via the two-step media API ([a81ab95](https://github.com/0xranx/golembot/commit/a81ab9566b5a233db81939ae9037f1ee3e834cee)), closes [#39](https://github.com/0xranx/golembot/issues/39)
+
+## [0.48.1](https://github.com/0xranx/golembot/compare/v0.48.0...v0.48.1) (2026-07-07)
+
+
+### Bug Fixes
+
+* **feishu:** enable websocket ping watchdog ([0fbc9be](https://github.com/0xranx/golembot/commit/0fbc9be1174923e1bb728154f2f6f60c90bc6505))
+
+# [0.48.0](https://github.com/0xranx/golembot/compare/v0.47.1...v0.48.0) (2026-06-12)
+
+
+### Features
+
+* **gateway:** add [CONTINUE] auto-continue relay for IM ([fe2ffd3](https://github.com/0xranx/golembot/commit/fe2ffd3d1596d4698f5f3337688b893bc49815bb)), closes [#37](https://github.com/0xranx/golembot/issues/37)
+
 ## [0.47.1](https://github.com/0xranx/golembot/compare/v0.47.0...v0.47.1) (2026-05-09)